refactore complete onboarding structure
